Make sure we don't keep a reference to the ZipPath in ArchivePathTree
A Path can actually contain references to the filesystem: this is the
case of the ZipPath.
We want to make sure we don't keep ZipFileSystem around when a path tree
is closed so we also need to nullify the path.

Finally, we make sure only DirectoryPathTree is marked as Serializable
as we only serialize the workspace directories.

(cherry picked from commit 6780c2c17788d59e30ff50238224416cd62786bb)
